backend/requirements.txt
Flask==3.0.0
cryptography==41.0.3
Pillow==10.0.0
